Changed all array params, fixed some overloads/overrides, Moved CRC32 out of array

+++ src/SoyArray.cpp
@@ -4,7 +4,7 @@
 #include "heaparray.hpp"
 #include "RemoteArray.h"
 
-std::string	SoyArray::OnCheckBoundsError(int Index,size_t Size,const std::string& Typename)
+std::string	SoyArray::OnCheckBoundsError(size_t Index,size_t Size,const std::string& Typename)
 {
 	std::stringstream Error;
 	Error << "Array<" << Typename << "> Index " << Index << "/" << Size << " out of bounds";
